% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scenario_slots.R
\name{mod_scenario_banner_ui}
\alias{mod_scenario_banner_ui}
\title{"Viewing saved scenario" banner}
\usage{
mod_scenario_banner_ui(id)
}
\arguments{
\item{id}{Module namespace ID -- must match the corresponding
\code{\link[=mod_scenario_slots_server]{mod_scenario_slots_server()}} call.}
}
\description{
"Viewing saved scenario" banner
}
